+++ lib/anyway/rails/loaders/yaml.rb
@@ -8,7 +8,11 @@ module Anyway
           parsed_yml = super
           return parsed_yml unless environmental?(parsed_yml)
 
-          super[::Rails.env] || {}
+          env_config = parsed_yml[::Rails.env] || {}
+          return env_config if Anyway::Settings.default_environmental_key.blank?
+
+          default_config = parsed_yml[Anyway::Settings.default_environmental_key] || {}
+          Utils.deep_merge!(default_config, env_config)
         end
 
         private
@@ -18,7 +22,9 @@ module Anyway
           # likely
           return true if parsed_yml.key?(::Rails.env)
           # less likely
-          ::Rails.application.config.anyway_config.known_environments.any? { parsed_yml.key?(_1) }
+          return true if ::Rails.application.config.anyway_config.known_environments.any? { parsed_yml.key?(_1) }
+          # strange, but still possible
+          Anyway::Settings.default_environmental_key.present? && parsed_yml.key?(Anyway::Settings.default_environmental_key)
         end
 
         def relative_config_path(path)